+++ src/ocean/text/xml/c/LibXml2.d
@@ -10,7 +10,7 @@
 
 *******************************************************************************/
 
-module ocean.text.xml.c.LibXml2;
+deprecated module ocean.text.xml.c.LibXml2;
 
 
 
@@ -215,6 +215,7 @@ extern ( C )
 
 *******************************************************************************/
 
+deprecated("This function is to be moved into another repository")
 public void formatXmlErrorString ( xmlErrorPtr err, ref mstring str )
 {
     cstring dstr ( char* cstr )